create fragment_home_admin + fragment_home_entrant + minor XML changes
Seperated fragment_home to fragment_home_admin + fragment_home_entrant + added butter title font and added padding to organizer options
