+++ app/src/main/java/com/example/butter/MainActivity.java
@@ -33,7 +33,7 @@ import java.util.Objects;
  * This activity is used to implement the navigation bar of the app with menu icons that
  * will have constraints depending on a users privileges.
  * User info is taken from firebase using a collection reference to the user document.
- * @author Angela
+ * @author Angela Dakay (angelcache)
  */
 public class MainActivity extends AppCompatActivity {
     /**
