+++ app/src/main/res/layout/activity_splash_screen.xml
@@ -14,7 +14,8 @@
         android:gravity="center"
         android:text="@string/butterLogo"
         android:textColor="@color/secondaryPurpleColor"
-        android:textSize="40sp"
+        android:textSize="80sp"
+        android:fontFamily="@font/butter_title_font"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintLeft_toLeftOf="parent"
         app:layout_constraintRight_toRightOf="parent"
